// frontend/contracts/core/v1/models/ContestJoinRequestModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ContestJoinRequestModel = {
    id: string;
    contest_id: string;
    contest_title: string;
    contest_login: string;
    organization_login: string;
    user_id: string;
    username: string;
    email?: string;
    message?: string;
    status: string;
    reviewer_username?: string;
    created_at: string;
    updated_at: string;
};