// contracts/core/v1/models/ContestMemberModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ContestMemberModel = {
    user_id: string;
    contest_id: string;
    username: string;
    role: string;
    contest_role: string;
    kratos_id: string;
    created_at: string;
    updated_at: string;
};